add about page by correcting the reviews
