+++ aait/web/group-4/starter-project/app/page.tsx
@@ -1,7 +1,9 @@
+import { About } from "@/components/about/About";
+
 export default function Home() {
   return (
     <main>
-     
+     <About/>
     </main>
   )
 }
